feat(projen.project.nx-monorepo): move organization common defaults to nx-monorepo project.
